+++ lib/models/account_type.dart
@@ -1,17 +1,17 @@
 class AccountType {
   static String materialFamily = "MaterialIcons";
-  late String _id = "GENERIC_INFORMATION";
+  late String _id;
   String name;
   int iconPoint;
   String iconFamily;
 
   AccountType({
-    String? id,
+    required String id,
     required this.name,
     required this.iconPoint,
     required this.iconFamily,
   }) {
-    _id = id ?? _id;
+    _id = id;
   }
 
   String get id => _id;
